+++ Core/Libraries/Source/WWVegas/WWLib/thread.h
@@ -18,6 +18,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 #include "always.h"
 #include "Vector.h"
 
@@ -91,6 +93,6 @@ protected:
 
 private:
 	static void __cdecl Internal_Thread_Function(void*);
-	volatile unsigned long handle;
+	volatile uintptr_t handle;
 	int thread_priority;
 };
